+++ typesafe_firebase_firestore/lib/typesafe_firebase_firestore.dart
@@ -1,3 +1,5 @@
 // Forestore
 
-export 'package:typesafe_firebase_firestore/firestore/base.dart' show $FirestoreDb, $Collection, $Document;
+export 'package:typesafe_firebase_firestore/firestore/collection.dart' show $Collection;
+export 'package:typesafe_firebase_firestore/firestore/document.dart' show $Document;
+export 'package:typesafe_firebase_firestore/firestore/base.dart' show $FirestoreDb;