+++ typesafe_firebase_firestore/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+
+* Support for getAll() for firestore collections.
+
 ## 0.0.2
 
 Basic firebase functionality and Examples.